Support Apple's universal binaries...
Autoconf will default to the underlying architecture, while distutils will try to build for everything the python binary supports, which leads to odd behavior.

Solution: Detect if we are on Darwin and need to build a universal binary (only
support i386 and x86_64).  If so, let autoconf know, so it can build Jerasure,
gf-complete and Xorcode as universal binaries.