Enable UnitTest to choose tests to do
This patch introduces a global variable "_available_backends"
meaning which backends are available in the test running environment.

PyECLib supports several backends and some of them needs extra
installation in outside of PyECLib setuptools. When such backends
doesn't be installed yet, some unit tests might fail.

To clarify which test should be done in the user environment and
to skip unnecessary tests if the backends are not available, we could
use the new variable with skipif decorator of unittest library.

E.g.:

from unittest import skipIf

@skipIf(PyECLib_EC_Types.shss not in _available_backends,
        "shss backend is not available in your enviromnet")
def test_shss(self):
    pass