Add close/destroy methods; clean up descriptors
Previously, we would *never* call liberasurecode_instance_destroy for
any call to liberasurecode_instance_create. That was bad and would lead
to memory leaks -- while the capsule destructor would eventually free
the handle it wrapped, it would do nothing to clean up the associated
liberasurecode descriptor.

Now, use liberasurecode_instance_destroy to attempt to clean up the
descriptor. This can happen in the destructor (in which case any errors
are suppressed), or in a new `close` method on both ECPyECLibDriver and
ECDriver. If attempts are made to use a driver after close,
ECBackendInstanceNotAvailable will be raised, as the descriptor has
already been destroyed. The capsule destructor continues to be
responsible for freeing the handle.

Two new error modes are also handled in `pyeclib_c_seterr`, as they may
arise from liberasurecode_instance_destroy -- EDEADLK and EINVAL will
raise an ECDriverError, indicating trouble with the global instance lock.

Co-Authored-By: Tim Burke <tim.burke@gmail.com>
Closes-Bug: #1954352
Change-Id: I460014871a88fd7439efce3c5e8333c8447d1bdd