wheels: Disable optimizations for liberasurecode
Previously we could build with, for example, AVX optimizations which
would not necessarily be available on all the machines we'd hope to be
able to use wheels.

Note that ISA-L has some fancy multi-binary-dispatchers thing going on
that should do the right thing at run-time, but liberasurecode doesn't
know how to do that.

Change-Id: Iadfe6e7c7820761040d0da278636350fc0b50f9d